reorganize components within DSL extensions
