+++ packages/legend-extension-dsl-data-space/package.json
@@ -39,6 +39,7 @@
   "dependencies": {
     "@finos/legend-application": "workspace:*",
     "@finos/legend-art": "workspace:*",
+    "@finos/legend-extension-dsl-diagram": "workspace:*",
     "@finos/legend-graph": "workspace:*",
     "@finos/legend-model-storage": "workspace:*",
     "@finos/legend-shared": "workspace:*",
