+++ packages/legend-extension-dsl-text/src/components/studio/DSLText_StudioPlugin.tsx
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import packageJson from '../../package.json';
+import packageJson from '../../../package.json';
 import { StudioPlugin } from '@finos/legend-studio';
 import type {
   StudioPluginManager,
@@ -30,10 +30,10 @@ import type {
   NewElementState,
 } from '@finos/legend-studio';
 import { FileIcon } from '@finos/legend-art';
-import { TextEditorState } from '../stores/TextEditorState';
+import { TextEditorState } from '../../stores/studio/TextEditorState';
 import { TextElementEditor } from './TextElementEditor';
 import type { PackageableElement } from '@finos/legend-graph';
-import { Text } from '../models/metamodels/pure/model/packageableElements/text/Text';
+import { Text } from '../../models/metamodels/pure/model/packageableElements/text/Text';
 
 const TEXT_ELEMENT_TYPE = 'TEXT';
 const TEXT_ELEMENT_PROJECT_EXPLORER_DND_TYPE = 'PROJECT_EXPLORER_TEXT';
