+++ packages/legend-extension-dsl-text/src/stores/studio/TextEditorState.ts
@@ -19,7 +19,7 @@ import type { EditorStore } from '@finos/legend-studio';
 import { guaranteeType } from '@finos/legend-shared';
 import { ElementEditorState } from '@finos/legend-studio';
 import type { PackageableElement } from '@finos/legend-graph';
-import { Text } from '../models/metamodels/pure/model/packageableElements/text/Text';
+import { Text } from '../../models/metamodels/pure/model/packageableElements/text/Text';
 
 export class TextEditorState extends ElementEditorState {
   constructor(editorStore: EditorStore, element: PackageableElement) {
